removed unsubscribed and subscribed events.
those events are misleading as there are no such presence packets
in the new rfc anymore about them. and they are not neccessary anymore
in the new spec. not even the 'unsubscribe' is delivered to the
clients anymore.

use the roster to track subscription states.

presence information is also detached from subscriptions as it
seems.